refactor: Split constants into networks and tokens
In advance of importing additional chain information from UMA.
