+++ src/networks.ts
@@ -1,0 +1,22 @@
+// Chain names and ID's
+export const CHAIN_IDs = {
+  MAINNET: 1,
+  OPTIMISM: 10,
+  POLYGON: 137,
+  BOBA: 288,
+  ZK_SYNC: 324,
+  BASE: 8453,
+  ARBITRUM: 42161,
+  GOERLI: 5,
+  OPTIMISM_GOERLI: 420,
+  ZK_SYNC_GOERLI: 280,
+  MUMBAI: 80001,
+  BASE_GOERLI: 84531,
+  ARBITRUM_GOERLI: 421613,
+  SEPOLIA: 11155111,
+  OPTIMISM_SEPOLIA: 11155420,
+  ZK_SYNC_SEPOLIA: 300,
+  POLYGON_AMOY: 80002,
+  BASE_SEPOLIA: 84532,
+  ARBITRUM_SEPOLIA: 421614,
+};
